pyproject.toml
[project]
name = "greenwashing-detective-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==6.2.4",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloud-sql-python-connector==1.19.0",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cryptography==46.0.3",
    "cycler==0.12.1",
    "dnspython==2.8.0",
    "flask==3.1.2",
    "flask-cors==6.0.2",
    "fonttools==4.61.1",
    "frozenlist==1.8.0",
    "gnews>=0.4.2",
    "google-auth==2.45.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "jieba==0.42.1",
    "jinja2==3.1.6",
    "kiwisolver==1.4.9",
    "line-bot-sdk>=3.21.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "multidict==6.7.0",
    "numpy==2.4.0",
    "packaging==25.0",
    "pandas>=2.3.3",
    "pdfminer-six==20251107",
    "pdfplumber==0.11.8",
    "perplexityai>=0.22.3",
    "pillow==12.0.0",
    "propcache==0.4.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pymysql==1.1.2",
    "pyparsing==3.3.1",
    "pypdfium2==5.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "six==1.17.0",
    "tiktoken>=0.12.0",
    "typing-extensions==4.15.0",
    "urllib3==2.6.2",
    "werkzeug==3.1.4",
    "wordcloud==1.9.5",
    "yarl==1.22.0",
]
